+++ src/main/java/com/company/project/api/CategoryController.java
@@ -1,18 +1,20 @@
 package com.company.project.web;
-
 import com.company.project.core.Result;
 import com.company.project.core.ResultGenerator;
 import com.company.project.model.Category;
 import com.company.project.service.CategoryService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
 import java.util.List;
 
 /**
-* Created by ado on 2018-09-09.
+* Created by ado on 2018-09-10.
 */
 @RestController
 @RequestMapping("/category")
@@ -20,35 +22,58 @@ public class CategoryController {
     @Resource
     private CategoryService categoryService;
 
-    @PostMapping
-    public Result add(@RequestBody Category category) {
+    @PostMapping("/add")
+    public Result add(Category category) {
         categoryService.save(category);
         return ResultGenerator.genSuccessResult();
     }
 
-    @DeleteMapping("/{id}")
-    public Result delete(@PathVariable Integer id) {
-        categoryService.deleteById(id);
-        return ResultGenerator.genSuccessResult();
+    @PostMapping("/delete")
+    public Result delete(@RequestParam(value = "id", required = true) Integer id) {
+        Category category = categoryService.findById(id);
+        if(category != null)
+        {
+            categoryService.deleteById(id);
+            return ResultGenerator.genSuccessResult();
+        }
+        else
+        {
+             return ResultGenerator.genFailResult("数据不存在，删除数据失败");
+        }
     }
 
-    @PutMapping
-    public Result update(@RequestBody Category category) {
+    @PostMapping("/update")
+    public Result update(Category category) {
         categoryService.update(category);
         return ResultGenerator.genSuccessResult();
     }
 
-    @GetMapping("/{id}")
-    public Result detail(@PathVariable Integer id) {
+    @PostMapping("/detail")
+    public Result detail(@RequestParam(value = "id", required = true) Integer id) {
         Category category = categoryService.findById(id);
-        return ResultGenerator.genSuccessResult(category);
+        if(category != null)
+        {
+            return ResultGenerator.genSuccessResult(category);
+        }
+        else
+        {
+            return ResultGenerator.genFailResult("数据不存在，获取详情失败");
+        }
     }
 
-    @GetMapping
+
+
+    @PostMapping("/list")
     public Result list(@RequestParam(defaultValue = "0") Integer page, @RequestParam(defaultValue = "0") Integer size) {
         PageHelper.startPage(page, size);
         List<Category> list = categoryService.findAll();
         PageInfo pageInfo = new PageInfo(list);
         return ResultGenerator.genSuccessResult(pageInfo);
     }
+
+
+
+
+
+
 }
