+++ src/main/java/com/company/project/api/RolesUserController.java
@@ -1,18 +1,20 @@
 package com.company.project.web;
-
 import com.company.project.core.Result;
 import com.company.project.core.ResultGenerator;
 import com.company.project.model.RolesUser;
 import com.company.project.service.RolesUserService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
 import java.util.List;
 
 /**
-* Created by ado on 2018-09-09.
+* Created by ado on 2018-09-10.
 */
 @RestController
 @RequestMapping("/roles/user")
@@ -20,35 +22,58 @@ public class RolesUserController {
     @Resource
     private RolesUserService rolesUserService;
 
-    @PostMapping
-    public Result add(@RequestBody RolesUser rolesUser) {
+    @PostMapping("/add")
+    public Result add(RolesUser rolesUser) {
         rolesUserService.save(rolesUser);
         return ResultGenerator.genSuccessResult();
     }
 
-    @DeleteMapping("/{id}")
-    public Result delete(@PathVariable Integer id) {
-        rolesUserService.deleteById(id);
-        return ResultGenerator.genSuccessResult();
+    @PostMapping("/delete")
+    public Result delete(@RequestParam(value = "id", required = true) Integer id) {
+        RolesUser rolesUser = rolesUserService.findById(id);
+        if(rolesUser != null)
+        {
+            rolesUserService.deleteById(id);
+            return ResultGenerator.genSuccessResult();
+        }
+        else
+        {
+             return ResultGenerator.genFailResult("数据不存在，删除数据失败");
+        }
     }
 
-    @PutMapping
-    public Result update(@RequestBody RolesUser rolesUser) {
+    @PostMapping("/update")
+    public Result update(RolesUser rolesUser) {
         rolesUserService.update(rolesUser);
         return ResultGenerator.genSuccessResult();
     }
 
-    @GetMapping("/{id}")
-    public Result detail(@PathVariable Integer id) {
+    @PostMapping("/detail")
+    public Result detail(@RequestParam(value = "id", required = true) Integer id) {
         RolesUser rolesUser = rolesUserService.findById(id);
-        return ResultGenerator.genSuccessResult(rolesUser);
+        if(rolesUser != null)
+        {
+            return ResultGenerator.genSuccessResult(rolesUser);
+        }
+        else
+        {
+            return ResultGenerator.genFailResult("数据不存在，获取详情失败");
+        }
     }
 
-    @GetMapping
+
+
+    @PostMapping("/list")
     public Result list(@RequestParam(defaultValue = "0") Integer page, @RequestParam(defaultValue = "0") Integer size) {
         PageHelper.startPage(page, size);
         List<RolesUser> list = rolesUserService.findAll();
         PageInfo pageInfo = new PageInfo(list);
         return ResultGenerator.genSuccessResult(pageInfo);
     }
+
+
+
+
+
+
 }
